Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #11:

This patch reduces the compiler warning count on Win64 by an additional
80, as reported by VS2013. The patch also removes a number of unnecessary casts.

+++ sql/handler.h
@@ -2889,7 +2889,7 @@ public:
   */
 
   virtual my_bool register_query_cache_table(THD *thd, char *table_key,
-                                             uint key_length,
+                                             size_t key_length,
                                              qc_engine_callback
                                              *engine_callback,
                                              ulonglong *engine_data)
